peashooter enemy detection - very big issue finally solved
Completely changed my code for my peashooter detection. used AI to suggest the use of an array and add all zombies to it and check each to see if it was in that lane but thats it. peashooter now fires but the projectile doesn't move yet. will be a job for the future.

+++ Assets/Scripts/Plant Scripts/ClassicPlantStats.cs
@@ -5,7 +5,7 @@ public class ClassicPlantStats : MonoBehaviour
 {
 
     public int maxHealth = 100;
-    public double fireCooldown = 2;
+    public float fireCooldown = 2f;
     public int currentHealth;
     private void Awake()
     {
